requirements.txt
absl-py==0.9.0
asn1crypto==1.3.0
astor==0.8.0
astroid==2.3.3
attrs==19.3.0
autokeras==1.0.2
backcall==0.2.0
beautifulsoup4==4.8.2
bleach==3.1.5
blinker==1.4
blis==0.4.1
Bottleneck==1.3.2
bunny==0.3
cachetools==4.0.0
catalogue==1.0.0
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.1
colorama==0.4.3
contextlib2==0.6.0.post1
cryptography==2.8
cupy-cuda100==7.3.0
cycler==0.10.0
cymem==2.0.3
Cython==0.29.20
dataclasses==0.7
decorator==4.4.2
defusedxml==0.6.0
entrypoints==0.3
et-xmlfile==1.0.1
fastai==1.0.60
fastprogress==0.2.2
fastrlock==0.4
future==0.18.2
gast==0.2.2
gitdb==4.0.4
GitPython==3.1.1
google-auth==1.11.2
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
grpcio==1.27.2
h5py==2.10.0
hottbox==0.3.2
idna==2.9
importlib-metadata==1.6.0
ipykernel==5.3.0
ipython==7.15.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
jdcal==1.4.1
jedi==0.17.0
Jinja2==2.11.2
joblib==0.14.1
json5==0.9.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.3
jupyter-console==6.1.0
jupyter-core==4.6.3
jupyter-kite==1.0.0
jupyterlab==2.1.0
jupyterlab-git==0.10.0
jupyterlab-server==1.1.1
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keras-tuner==1.0.1
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
lxml==4.5.1
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.2.1
mccabe==0.6.1
mistune==0.8.4
mkl-fft==1.0.15
mkl-random==1.1.0
mkl-service==2.3.0
mpmath==1.1.0
murmurhash==1.0.2
nbconvert==5.6.1
nbdime==1.1.0
nbformat==5.0.7
notebook==6.0.3
np-utils==0.5.12.1
numexpr==2.7.1
numpy==1.18.1
nvidia-ml-py3==7.352.0
oauthlib==3.1.0
opencv-python==4.2.0.34
openpyxl==3.0.3
opt-einsum==3.1.0
packaging==20.4
pandas==1.0.3
pandocfilters==1.4.2
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.0.0
plac==1.1.3
preshed==3.0.2
prometheus-client==0.8.0
prompt-toolkit==3.0.5
protobuf==3.11.4
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.6.1
PyJWT==1.7.1
pylint==2.4.4
pyOpenSSL==19.1.0
pyparsing==2.4.7
pyrsistent==0.16.0
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3
pyzmq==19.0.1
qtconsole==4.7.4
QtPy==1.9.0
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
scikit-learn==0.22.1
scipy==1.4.1
seaborn==0.10.0
Send2Trash==1.5.0
six==1.14.0
smmap==3.0.2
soupsieve==2.0
spacy==2.2.3
srsly==1.0.1
sympy==1.5.1
tabulate==0.8.6
tensorboard==2.1.1
tensorflow-estimator==2.1.0
tensorflow-gpu==2.1.0
termcolor==1.1.0
terminado==0.8.3
terminaltables==3.1.0
testpath==0.4.4
tf-slim==1.1.0
thinc==7.3.1
torch==1.4.0
torchvision==0.5.0
tornado==6.0.4
tqdm==4.42.1
traitlets==4.3.3
typed-ast==1.4.1
urllib3==1.25.8
wasabi==0.6.0
wcwidth==0.2.4
webencodings==0.5.1
Werkzeug==1.0.0
widgetsnbextension==3.5.1
wrapt==1.12.1
xlrd==1.2.0
zipp==3.1.0
